backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-openai
chromadb
python-dotenv
PyPDF2
pytest
pypdf
python-multipart
celery[redis]>=5.3
redis>=5.0
langchain-chroma
chromadb